Move share search function to `lib` folder
